up to date

+++ userbot/plugins/help.py
@@ -3,17 +3,19 @@ import asyncio
 import requests
 from telethon import functions
 
-from . import CMD_LIST, SUDO_LIST, mention, yaml_format
-
-HELPTYPE = Config.HELP_INLINETYPE or True
+from . import ALIVE_NAME, CMD_LIST, SUDO_LIST
+from .sql_helper.globals import addgvar, gvarstatus
 
 
 @bot.on(admin_cmd(outgoing=True, pattern="help ?(.*)"))
 async def cmd_list(event):
-    global HELPTYPE
-    reply_to_id = None
-    if event.reply_to_msg_id:
-        reply_to_id = event.reply_to_msg_id
+    if event.fwd_from:
+        return
+    if gvarstatus("HELPTYPE") and gvarstatus("HELPTYPE") == "false":
+        HELPTYPE = False
+    else:
+        HELPTYPE = True
+    reply_to_id = await reply_id(event)
     input_str = event.pattern_match.group(1)
     if input_str == "text":
         string = (
@@ -62,22 +64,20 @@ async def cmd_list(event):
             await event.delete()
     else:
         if HELPTYPE is True:
-            help_string = f"Userbot Helper. Provided by {mention} to reveal all the plugins\
+            help_string = f"Userbot Helper. Provided by {ALIVE_NAME} to reveal all the plugins\
                           \nCheck `.help plugin name` for commands, in case popup doesn't appear.\
                           \nCheck `.info plugin name` for usage of thoose plugins and commands"
-            tgbotusername = Config.TG_BOT_USER_NAME_BF_HER
-            results = await bot.inline_query(  # pylint:disable=E0602
-                tgbotusername, help_string
-            )
+            tgbotusername = Config.TG_BOT_USERNAME
+            results = await event.client.inline_query(tgbotusername, help_string)
             await results[0].click(event.chat_id, reply_to=reply_to_id, hide_via=True)
             await event.delete()
         else:
-            string = "<b>Harap tentukan fitur mana yang Anda ingin lihat!!\
-                \nJumlah fitur : </b><code>{count}</code>\
-                \n<b>Pemakaian:</b> <code>.info nama fitur</code> \n\n"
+            string = "<b>Please specify which plugin do you want help for !!\
+                \nNumber of plugins : </b><code>{count}</code>\
+                \n<b>Usage:</b> <code>.help plugin name</code> \n\n"
             catcount = 0
             for i in sorted(CMD_LIST):
-                string += "•" + f"<code>{str(i)}</code>"
+                string += "◆ " + f"<code>{str(i)}</code>"
                 string += " "
                 catcount += 1
             await event.edit(string.format(count=catcount), parse_mode="HTML")
@@ -85,14 +85,16 @@ async def cmd_list(event):
 
 @bot.on(sudo_cmd(allow_sudo=True, pattern="help ?(.*)"))
 async def info(event):
+    if event.fwd_from:
+        return
     input_str = event.pattern_match.group(1)
     if input_str == "text":
-        string = "jumlah {count} Fitur adalah{plugincount} Fitur sudo PRINDAPANBOT\n\n"
+        string = "Total {count} commands found in {plugincount} sudo plugins of catuserbot\n\n"
         catcount = 0
         plugincount = 0
         for i in sorted(SUDO_LIST):
             plugincount += 1
-            string += f"{plugincount}) Perintah menggunakan fitur " + i + " are \n"
+            string += f"{plugincount}) Commands found in Plugin " + i + " are \n"
             for iter_list in SUDO_LIST[i]:
                 string += "    " + str(iter_list)
                 string += "\n"
@@ -109,7 +111,7 @@ async def info(event):
                 .get("key")
             )
             url = f"https://nekobin.com/{key}"
-            reply_text = f"Lihat lengkap jumlah fitur[here]({url})"
+            reply_text = f"All commands of the catuserbot are [here]({url})"
             await event.reply(reply_text, link_preview=False)
             return
         await event.reply(
@@ -128,17 +130,17 @@ async def info(event):
                 string.format(count=catcount, input_str=input_str), parse_mode="HTML"
             )
         else:
-            reply = await event.reply(input_str + " Bukan plugin yang benar!")
+            reply = await event.reply(input_str + " is not a valid plugin!")
             await asyncio.sleep(3)
             await event.delete()
             await reply.delete()
     else:
-        string = "<b>Harap tentukan fitur mana yang Anda inginkan bantuannya !!\
-            \Jumlah Fitur : </b><code>{count}</code>\
-            \n<b>Pemakaian:</b> <code>.help nama fitur</code>\n\n"
+        string = "<b>Please specify which plugin do you want help for !!\
+            \nNumber of plugins : </b><code>{count}</code>\
+            \n<b>Usage:</b> <code>.help plugin name</code>\n\n"
         catcount = 0
         for i in sorted(SUDO_LIST):
-            string += "• " + f"<code>{str(i)}</code>"
+            string += "◆ " + f"<code>{str(i)}</code>"
             string += " "
             catcount += 1
         await event.reply(string.format(count=catcount), parse_mode="HTML")
@@ -147,19 +149,20 @@ async def info(event):
 @bot.on(admin_cmd(outgoing=True, pattern="info ?(.*)"))
 @bot.on(sudo_cmd(pattern="info ?(.*)", allow_sudo=True))
 async def info(event):
-    """ For .info command,"""
+    if event.fwd_from:
+        return
     args = event.pattern_match.group(1).lower()
     if args:
         if args in CMD_HELP:
             await edit_or_reply(event, str(CMD_HELP[args]))
         else:
-            event = await edit_or_reply(event, "Punya otak? Fitur tidak ada gblk.")
+            event = await edit_or_reply(event, "Please specify a valid plugin name.")
             await asyncio.sleep(3)
             await event.delete()
     else:
-        string = "<b>Harap tentukan plugin mana yang Anda inginkan bantuannya!\
-            \nJumlah fitur : </b><code>{count}</code>\
-            \n<b>Pemakaian : </b><code>.info nama fitur</code>\n\n"
+        string = "<b>Please specify which plugin do you want help for !!\
+            \nNumber of plugins : </b><code>{count}</code>\
+            \n<b>Usage : </b><code>.info plugin name</code>\n\n"
         catcount = 0
         for i in sorted(CMD_HELP):
             string += "◆ " + f"<code>{str(i)}</code>"
@@ -176,9 +179,9 @@ async def info(event):
 async def _(event):
     if event.fwd_from:
         return
-    result = await bot(functions.help.GetNearestDcRequest())
+    result = await event.client(functions.help.GetNearestDcRequest())
     result = (
-        yaml_format(result)
+        _format.yaml_format(result)
         + "\n\n**List Of Telegram Data Centres:**\
                 \nDC1 : Miami FL, USA\
                 \nDC2 : Amsterdam, NL\
@@ -194,21 +197,21 @@ async def _(event):
 async def _(event):
     if event.fwd_from:
         return
-    global HELPTYPE
     input_str = event.pattern_match.group(1)
-    if input_str == "true":
-        type = True
+    h_type = input_str == "true"
+    if gvarstatus("HELPTYPE") and gvarstatus("HELPTYPE") == "false":
+        HELPTYPE = False
     else:
-        type = False
-    if HELPTYPE is True:
-        if type is True:
+        HELPTYPE = True
+    if HELPTYPE:
+        if h_type:
             await event.edit("`inline mode is already enabled`")
         else:
-            HELPTYPE = type
+            addgvar("HELPTYPE", h_type)
             await event.edit("`inline mode is disabled`")
     else:
-        if type is True:
-            HELPTYPE = type
+        if h_type:
+            addgvar("HELPTYPE", h_type)
             await event.edit("`inline mode is enabled`")
         else:
             await event.edit("`inline mode is already disabled`")
@@ -216,14 +219,18 @@ async def _(event):
 
 CMD_HELP.update(
     {
-        "help": "__**PLUGIN NAME :** Help__\
-    \n\n📌** CMD ➥** `.help` <plugin name>\
-    \n**USAGE   ➥  **Get commands of a specific plugin.\
-    \n\n📌** CMD ➥** `.info` <plugin name>\
-    \n**USAGE   ➥  **To get commands and usage of a specific plugin.\
-    \n\n📌** CMD ➥** `.dc`\
-    \n**USAGE   ➥  **To get info list of telegram data centres.\
-    \n\n📌** CMD ➥** `.setinline true` <or> `.setinline false`\
-    \n**USAGE   ➥  **To enable or disable inline mode of help menu."
+        "help": """**Plugin : **`help`
+
+•  **Syntax : **`.help/.help plugin_name`
+•  **Function : **__If you just type .help then shows you help menu, if plugin name is given then shows you only commands in thst plugin and if you use `.help text` then shows you all commands in your userbot__
+
+•  **Syntax : **`.info/.info plugin_name`
+•  **Function : **__To get details/information/usage of that plugin__
+
+•  **Syntax : **`.dc`
+•  **Function : **__Shows your dc id and dc ids list__
+
+•  **Syntax : **`.setinline (true|false)`
+•  **Function : **__Sets help menu either in inline or text format__"""
     }
 )
